requirements.txt
h5py==2.10.0
scipy==1.4.1
numpy
Keras==2.4.3
streamlit==0.79.0
Pillow==8.1.2
tensorflow==2.3.0
